+++ tests/0030_manifest/0032_dump.py
@@ -22,6 +22,7 @@ def test_manifest_to_dict():
         "}\n"
         "\n"
         ".styleguide-reference-text_color{\n"
+        '    --structure: "nested";\n'
         '    --keys: "black white";\n'
         '    --selectors: ".bg-black .bg-white";\n'
         '    --values: "#000000 #ffffff";\n'
@@ -66,6 +67,7 @@ def test_manifest_to_json():
         "}\n"
         "\n"
         ".styleguide-reference-text_color{\n"
+        '    --structure: "nested";\n'
         '    --keys: "black white";\n'
         '    --selectors: ".bg-black .bg-white";\n'
         '    --values: "#000000 #ffffff";\n'